fix: match cookbook structure exactly with src/ directory
- Move main.py to src/main.py
- Add src/middleware.py with SmitheryConfigMiddleware
- Update Dockerfile to run src/main.py
- Add back TRANSPORT=http, PORT=8080, EXPOSE 8080

Exactly matches the working cookbook server structure at:
https://github.com/smithery-ai/cookbook/tree/main/servers/python/migrate_stdio_to_http
