--- main.py
@@ -1,44 +1,0 @@
-"""DevPlan MCP Server - HTTP entry point for Smithery deployment."""
-
-import os
-import uvicorn
-from mcp.server.fastmcp import FastMCP
-from starlette.middleware.cors import CORSMiddleware
-
-# Initialize MCP server
-mcp = FastMCP(name="DevPlan")
-
-
-@mcp.tool()
-def hello(name: str) -> str:
-    """Say hello to someone."""
-    return f"Hello, {name}!"
-
-
-def main():
-    """Run the server in HTTP mode for Smithery."""
-    print("DevPlan MCP Server starting...")
-
-    # Setup Starlette app with CORS for cross-origin requests
-    app = mcp.streamable_http_app()
-
-    # Add CORS middleware for browser-based clients
-    app.add_middleware(
-        CORSMiddleware,
-        allow_origins=["*"],
-        allow_credentials=True,
-        allow_methods=["GET", "POST", "OPTIONS"],
-        allow_headers=["*"],
-        expose_headers=["mcp-session-id", "mcp-protocol-version"],
-        max_age=86400,
-    )
-
-    # Get port from environment variable (Smithery sets PORT)
-    port = int(os.environ.get("PORT", 8080))
-    print(f"Listening on port {port}")
-
-    uvicorn.run(app, host="0.0.0.0", port=port, log_level="debug")
-
-
-if __name__ == "__main__":
-    main()
